Move AppleScript MissingArtwork extension code into its own file
It was getting unwieldly.
